Put TEASER sim_settings into the plugin
