+++ bim2sim/sim_settings.py
@@ -9,7 +9,7 @@ from pathlib import Path
 from typing import Union
 
 from bim2sim.utilities import types
-from bim2sim.utilities.types import LOD, ZoningCriteria
+from bim2sim.utilities.types import LOD
 from bim2sim.elements.base_elements import Material
 from bim2sim.elements import bps_elements as bps_elements,\
     hvac_elements as hvac_elements
@@ -765,50 +765,6 @@ class LCAExportSettings(BuildingSimSettings):
 
 # TODO #511 Plugin specific sim_settings temporary needs to be stored here to
 #  prevent import problems during integration tests
-class TEASERSimSettings(BuildingSimSettings):
-    """Defines simulation settings for TEASER Plugin.
-
-    This class defines the simulation settings for the TEASER Plugin. It
-    inherits all choices from the BuildingSimulation settings. TEASER
-    specific settings are added here..
-    """
-
-    zoning_setup = ChoiceSetting(
-        default=LOD.low,
-        choices={
-            LOD.low: 'All IfcSpaces of the building will be merged into '
-                     'one thermal zone.',
-            LOD.medium: 'IfcSpaces of the building will be merged together'
-                        ' based on selected zoning criteria.',
-            LOD.full: 'Every IfcSpace will be a separate thermal zone'
-        },
-        description='Select the criteria based on which thermal zones will '
-                    'be aggreated.',
-        for_frontend=True
-    )
-
-    zoning_criteria = ChoiceSetting(
-        default=ZoningCriteria.usage,
-        choices={
-            ZoningCriteria.external:
-                'Group all thermal zones that have contact to the exterior'
-                ' together and all thermal zones that do not have contact to'
-                ' exterior.',
-            ZoningCriteria.external_orientation:
-                'Like external, but takes orientation '
-                '(North, east, south, west)'
-                ' into account as well',
-            ZoningCriteria.usage:
-                'Group all thermal zones that have the same usage.',
-            ZoningCriteria.external_orientation_usage:
-                'Combines the prior options.',
-            ZoningCriteria.all_criteria:
-                'Uses all prior options and adds glass percentage of the rooms'
-                ' as additional criteria and only groups rooms if they are'
-                ' adjacent to each other.'
-        },
-        for_frontend=True
-    )
 
 
 class EnergyPlusSimSettings(BuildingSimSettings):
