+++ test/unit/tasks/common/test_weather.py
@@ -6,7 +6,8 @@ import bim2sim.tasks.common.create_elements
 from bim2sim.kernel.decision.decisionhandler import DebugDecisionHandler
 from bim2sim.plugins import Plugin
 from bim2sim.project import Project
-from bim2sim.sim_settings import TEASERSimSettings, EnergyPlusSimSettings
+from bim2sim.sim_settings import EnergyPlusSimSettings
+from bim2sim_teaser import TEASERSimSettings
 from bim2sim.utilities.types import IFCDomain
 from bim2sim.tasks.common import Weather
